cleanup more warnings

+++ libraries/ESPmDNS/src/ESPmDNS.cpp
@@ -308,7 +308,7 @@ bool MDNSResponder::hasTxt(int idx, const char * key) {
         log_e("Result %d not found", idx);
         return false;
     }
-    int i = 0;
+    size_t i = 0;
     while(i < result->txt_count) {
         if (strcmp(result->txt[i].key, key) == 0) return true;
         i++;
@@ -322,7 +322,7 @@ String MDNSResponder::txt(int idx, const char * key) {
         log_e("Result %d not found", idx);
         return "";
     }
-    int i = 0;
+    size_t i = 0;
     while(i < result->txt_count) {
         if (strcmp(result->txt[i].key, key) == 0) return result->txt[i].value;
         i++;
@@ -336,7 +336,7 @@ String MDNSResponder::txt(int idx, int txtIdx) {
         log_e("Result %d not found", idx);
         return "";
     }
-    if (txtIdx >= result->txt_count) return "";
+    if ((unsigned int)txtIdx >= result->txt_count) return "";
     return result->txt[txtIdx].value;
 }
